+++ crates/action/src/errors.rs
@@ -1,4 +1,0 @@
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum ActionError {}