+++ crates/action/src/lib.rs
@@ -1,4 +1,3 @@
 mod action;
-mod errors;
 
 pub use action::*;