+++ packages/cli/CHANGELOG.md
@@ -9,10 +9,12 @@
   - To start, `moon query touched-files` can be used to query touched files. The same files
     `moon ci` and `moon run` use.
   - Also `moon query projects` can be used to query about projects in the project graph.
+- Added `bash` as a support value for the project `language` setting.
 
 #### ⚙️ Internal
 
 - Updated Rust to v1.62.
+- Refactored our action runner to support additional languages in the future.
 
 ## 0.5.0